config/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.3
Django==4.0.2
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-nine==0.2.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.2
filelock==3.3.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
pipenv==2021.5.29
platformdirs==2.4.0
pycosat==0.6.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
